data_collection/altdata_service/news/requirements.txt
pandas
pyaml
feedparser
numpy
textblob
psycopg2-binary
requests
sqlalchemy
bs4
langdetect
vaderSentiment
lxml
OpenPermID
scipy
#sec-edgar-downloader